+++ Changelog.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Added
 - Added [DAE](models/goyal2020/Readme.md)
+- Adding [FactCC and FactCCX](models/kryscinski2019/Readme.md)
 
 ## [v0.0.3](https://github.com/danieldeutsch/repro/releases/tag/v0.0.3) - 2021-08-04
 ### Added