+++ Changelog.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Added [DAE](models/goyal2020/Readme.md)
 - Adding [FactCC and FactCCX](models/kryscinski2019/Readme.md)
+- Added utilities to remove empty inputs and insert values at specific indices
+
+### Changed
+- BERTScore now returns 0 for its metrics if the input is empty. 
 
 ## [v0.0.3](https://github.com/danieldeutsch/repro/releases/tag/v0.0.3) - 2021-08-04
 ### Added