+++ Changelog.md
@@ -4,7 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [v0.1.2](https://github.com/danieldeutsch/repro/releases/tag/v0.1.1) - 2021-10-07
+### Changed
+- Changed backend of Lite3Pyramid to use our own fork of the official repo with some modifications.
 
 ## [v0.1.1](https://github.com/danieldeutsch/repro/releases/tag/v0.1.1) - 2021-10-05
 ### Added